requirements/requirements.txt
attrs==22.1.0
certifi==2022.5.18.1
iniconfig==1.1.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
tomli==2.0.1
